Refactor 01: breakdown envelope processor (#291)
* Move inner enum to separate file

* Move inner processing result class to separate file

* Move processing exception to different package
