+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/servicebus/handler/MessageProcessingResultType.java
@@ -1,0 +1,7 @@
+package uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.handler;
+
+public enum MessageProcessingResultType {
+    SUCCESS,
+    UNRECOVERABLE_FAILURE,
+    POTENTIALLY_RECOVERABLE_FAILURE
+}
